Revert "[#2] Caches the default registry"
Instead of embeding the cache into the JS files, we'll follow the
datapackage-registry-py pattern and load from the filesystem (when running in
NodeJS) or not have a cache (when in the browser).

This avoids issues with the schemas using `$ref` to point to other schemas, and
also will make the JS file smaller.

This reverts commit d4073c12fb9e11e47e44ed1d7c098e9b0fedb599.

Conflicts:
	lib/index.js
	test/test.js