[#2] Compile package to ES5 for NPM
The ./src folder contains our code (ES6), the ./lib contains it compiled to
ES5, and the schemas cache was moved from ./lib/schemas to ./schemas.

In our tests, I've prefered to test the compiled ES5, instead of our ES6 code.
This means that the stacktraces will point to the compiled ES5 code, but if
they pass, we can be sure that the resulting package works (including the
compilation process).